LayoutTests/fast/canvas/webgl/WebGLContextEvent-expected.txt

Test that the event passed to a listener of webglcontextlost is a WebGLContextEvent.
PASS evt.toString() is '[object WebGLContextEvent]'
PASS evt.statusMessage is ''
Test that the event passed to a listener of webglcontextrestored is a WebGLContextEvent.
PASS evt.toString() is '[object WebGLContextEvent]'
PASS evt.statusMessage is ''
PASS extension.restoreContext() was expected value: NO_ERROR.
